Allégement de loaders.js et séparation des claviers dans des chichiers différents
